+++ tests/testthat/test-testOutputs.R
@@ -6,7 +6,8 @@ test_that("Function response type matches expectation", {
 
   library(jsonlite)
 
-  #saveRDS(account_number,'/Users/tonytrevisan/Downloads/actnum.RDS')
+  # saveRDS(account_number,'/Users/tonytrevisan/Downloads/actnum.RDS')
+  # saveRDS(accessTokenList,'/Users/tonytrevisan/Downloads/actlist.RDS')
   account_number = readRDS('/Users/tonytrevisan/Downloads/actnum.RDS')
   accessTokenList = readRDS('/Users/tonytrevisan/Downloads/actlist.RDS')
   options(schwab_access_token = accessTokenList)