+++ NEWS.md
@@ -2,7 +2,12 @@
 
 ## Release Notes and News
 
-### schwabr 0.1.3 - 2/23/2025
+### schwabr 0.1.4 - 11/3/2025
+
+Adding parameters `range` for `schwab_optionChain`. Also adding function for 
+`schwab_replaceOrder`.
+
+### schwabr 0.1.3 - 2/23/2025 4c1a7ceeeaa9
 
 Adding user preferences API call.